Updated the ColorDetection library to actually read colors
Fixed some pin settings (in/output) and changed RGB struct variables
from to int.

+++ Libraries/ColorDetector/RGB.h
@@ -11,9 +11,9 @@
 
 
 struct RGB {
-	byte r;
-	byte g;
-	byte b;
+	int r;
+	int g;
+	int b;
 };
 
 #endif
